Rename SDK file to integration_sdk.js
